backend/requirements-dev.txt
# Coverage: generate code coverage reports
coverage==6.3.1

# Pylint: Python linter
pylint==2.12.2

# Pytest: Python testing framework
# (more advanced than the built-in unittest module)
pytest==7.0.1

# Pytest-cov: coverage plugin for pytest
pytest-cov==3.0.0

# Pytest-env: env plugin for pytest
pytest-env==0.6.2

# Pylint: check your codestyle
pylint==2.12.2

# Mypy: check type usage in code
mypy==0.931

# Sqlalchemy-stubs: type hints for sqlalchemy
sqlalchemy2-stubs==0.0.2a20

# Pylint-Pytest: A Pylint plugin to suppress pytest-related false positives.
pylint-pytest==1.1.2

# faker: Generate dummy data
faker==13.3.1

# Types for the passlib library
types-passlib==1.7.0

# Pytest-mock: mocking for pytest
pytest-mock==3.7.0
